+++ test/integration/query/count/test.bats
@@ -7,10 +7,10 @@ teardown() {
 }
 
 @test "Query: Number of Allocations per Aggregate" {
-    go test ./test/integration/query/count/allocations_per_aggregate_test.go
+    go test ./test/integration/query/count/allocations_running_pods_test.go
 }
 
 @test "Query: Number of Allocations Summary" {
-    go test ./test/integration/query/count/allocations_summary_test.go
+    go test ./test/integration/query/count/allocations_summary_running_pods_test.go
 }
